requirements.txt
streamlit
tensorflow
pillow
numpy
fpdf
yagmail
pandas
gdown
